reorg util
